adding module for hospital , doctor , Animal Guardian,

+++ veternary_website/src/main/java/com/ptpl/veternary_website/exceptions/ResourceNotFoundException.java
@@ -8,12 +8,20 @@ import lombok.Setter;
 public class ResourceNotFoundException extends RuntimeException {
 	String resourceName;
 	String fieldName;
-	Long fieldValue;
-	public ResourceNotFoundException(String resourceName, String fieldName, Long fieldValue) {
+	Integer fieldValue;
+//	String FieldStringValues;
+	public ResourceNotFoundException(String resourceName, String fieldName, Integer fieldValue) {
 		super(String.format("%s not found with %s : %s",resourceName,fieldName,fieldValue ));
 		this.resourceName = resourceName;
 		this.fieldName = fieldName;
 		this.fieldValue = fieldValue;
 	}
+	
+//	public ResourceNotFoundException(String resourceName, String fieldName, String fieldValue) {
+//		super(String.format("%s not found with %s : %s",resourceName,fieldName,fieldValue ));
+//		this.resourceName = resourceName;
+//		this.fieldName = fieldName;
+//		this.FieldStringValues = fieldValue;
+//	}
 	
 }
