+++ veternary_website/src/main/java/com/ptpl/veternary_website/services/UserService.java
@@ -6,11 +6,11 @@ import java.util.List;
 public interface UserService {
 	UserDto createUser(UserDto userDto);
 
-	UserDto updateUser(UserDto userDto, Long userId);
+	UserDto updateUser(UserDto userDto, Integer userId);
 
-	UserDto getUserById(Long userId);
+	UserDto getUserById(Integer userId);
 
 	List<UserDto> getAllUsers();
 	
-	void deleteUserById(Long userId);
+	void deleteUserById(Integer userId);
 }
